(feat) WIP

+++ hummingbot/strategy/pure_market_making/data_types.py
@@ -5,6 +5,8 @@ from typing import (
     List
 )
 
+from decimal import Decimal
+
 from hummingbot.core.event.events import OrderType
 from hummingbot.market.market_base import MarketBase
 
@@ -15,22 +17,22 @@ ORDER_PROPOSAL_ACTION_CANCEL_ORDERS = 1 << 1
 class OrdersProposal(NamedTuple):
     actions: int
     buy_order_type: OrderType
-    buy_order_prices: List[float]
-    buy_order_sizes: List[float]
+    buy_order_prices: List[Decimal]
+    buy_order_sizes: List[Decimal]
     sell_order_type: OrderType
-    sell_order_prices: List[float]
-    sell_order_sizes: List[float]
+    sell_order_prices: List[Decimal]
+    sell_order_sizes: List[Decimal]
     cancel_order_ids: List[str]
 
 
 class PricingProposal(NamedTuple):
-    buy_order_prices: List[float]
-    sell_order_prices: List[float]
+    buy_order_prices: List[Decimal]
+    sell_order_prices: List[Decimal]
 
 
 class SizingProposal(NamedTuple):
-    buy_order_sizes: List[float]
-    sell_order_sizes: List[float]
+    buy_order_sizes: List[Decimal]
+    sell_order_sizes: List[Decimal]
 
 
 class MarketInfo(NamedTuple):
